audio_vector_dataset_creator/requirements.txt
praat-parselmouth
pandas
chromadb
numpy